Fix critical auth module packaging issue - move to esologs.auth
