+++ docs/index.md
@@ -44,7 +44,7 @@
     ```python
     import asyncio
     from esologs.client import Client
-    from access_token import get_access_token
+    from esologs.auth import get_access_token
 
     async def main():
         token = get_access_token()
