+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "esologs-python"
-version = "0.2.0a1"
+version = "0.2.0a2"
 description = "A Python client library for the ESO Logs API v2"
 readme = "README.md"
 license = {text = "MIT"}
