+++ tests/docs/conftest.py
@@ -4,7 +4,7 @@ import os
 
 import pytest
 
-from access_token import get_access_token
+from esologs.auth import get_access_token
 
 
 @pytest.fixture(scope="session")
