+++ tests/docs/test_quickstart_examples.py
@@ -7,7 +7,7 @@ execute without errors and produce expected results.
 
 import pytest
 
-from access_token import get_access_token
+from esologs.auth import get_access_token
 from esologs.client import Client
 from esologs.exceptions import (
     GraphQLClientGraphQLError,
