+++ tests/integration/test_character_rankings.py
@@ -4,7 +4,7 @@ import asyncio
 
 import pytest
 
-from access_token import get_access_token
+from esologs.auth import get_access_token
 from esologs.client import Client
 from esologs.enums import CharacterRankingMetricType
