+++ tests/integration/test_core_api.py
@@ -4,7 +4,7 @@ import asyncio
 
 import pytest
 
-from access_token import get_access_token
+from esologs.auth import get_access_token
 from esologs.client import Client
 
 # Fixtures are now centralized in conftest.py
